+++ packages/admin-ui/src/lib/catalog/src/components/create-product-variant-dialog/create-product-variant-dialog.component.ts
@@ -67,7 +67,7 @@ export class CreateProductVariantDialogComponent implements Dialog<CreateProduct
 
     confirm() {
         const { name, sku, options, price } = this.form.value;
-        if (!name || !sku || !options || !price) {
+        if (!name || !sku || !options || price == null) {
             return;
         }