+++ packages/dev-server/test-plugins/experimental-ui/providers/cms-data.service.ts
@@ -1,0 +1,14 @@
+import { Injectable } from '@angular/core';
+import { of } from 'rxjs';
+
+@Injectable()
+export class CmsDataService {
+    getDataFor(id: string) {
+        return of({
+            id,
+            blogPostUrl: 'https://example.com/blog/' + id,
+            blogPostTitle: 'Example Blog Post',
+            blogTags: ['tag1', 'tag2', 'tag3'],
+        });
+    }
+}